Adjust perf tooling to not reference the crate name
