+++ tooling/perf/src/main.rs
@@ -46,7 +46,9 @@
 //! This should probably not be called manually unless you're working on the profiler
 //! itself; use the `cargo perf-test` alias (after building this crate) instead.
 
-use perf::{FailKind, Importance, Output, TestMdata, Timings, consts};
+mod implementation;
+
+use implementation::{FailKind, Importance, Output, TestMdata, Timings, consts};
 
 use std::{
     fs::OpenOptions,
